feat: Add InitializationProvider
