+++ src/app/lib/providers/index.ts
@@ -1,4 +1,5 @@
 export * from './AuthProvider';
+export * from './InitializationProvider';
 export * from './RecoilRootProvider';
 export * from './StyledComponentsRegistry';
 export * from './TanstackQueryProvider';
